Add test cases and Fix bugs

+++ imageservice/src/imageservice/routes/delete_image.py
@@ -1,5 +1,6 @@
+from decimal import Decimal
 from fastapi import APIRouter, Query,Request,Response
-from utils import Metadata, MetadataInputModel,ResponseModel
+from imageservice.utils import Metadata, MetadataInputModel,ResponseModel
 from starlette.responses import JSONResponse
 
 router = APIRouter()
@@ -7,7 +8,7 @@ router = APIRouter()
 @router.post("/delete",response_model=ResponseModel)
 def delete_image(request:Request,
                  userId: str = Query(default=None, description="Id of the User to delete images"),
-                 timestamp: int = Query(default=None, description="Timestamp of the image creation")):
+                 timestamp: Decimal = Query(default=None, description="Timestamp of the image creation")):
     """
     API used to delete image
     - **userId**: The userID of the item to view in query parameters
